refactor: split metadata.py into focused router modules
Split the 981-line monolithic metadata.py into a metadata/ package with
focused sub-modules:

- files.py: file/folder listing, browsing, filtering, CRUD, finalize
- artwork.py: artwork get/update/delete
- audio.py: peaks, streaming, transcoding
- collection.py: collection statistics
- proxy.py: SoundCloud CDN image proxy
- _helpers.py: shared resolve_folder utility

All 16 routes preserved. __init__.py combines sub-routers under
/api/metadata prefix.

Closes #42
